Resources now belongs to App context.
